requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
anyio==3.3.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
black==21.5b2
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fsspec==2021.6.1
future==0.18.2
google-auth==1.30.2
google-auth-oauthlib==0.4.1
grpcio==1.38.1
hydra-core==1.1.0
idna==2.10
importlib-metadata==4.5.0
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.0
packaging==20.9
pandocfilters==1.4.2
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pyDeprecate==0.3.0
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.3.7
pytz==2021.1
pyu2f==0.1.5
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.4.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.7.1
setuptools==52.0.0.post20210125
six==1.16.0
sniffio==1.2.0
tensorboard==2.4.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.3.2
torchvision==0.10.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.6.3
zipp==3.4.1
